# docs/openapi.yml
openapi: 3.0.3
info:
  title: NaHerbs Website MVP API
  version: 2.2.0
  description: 'API contract for naherb-api (Spring Boot) used by naherb-web (Next.js).
    Updated for UUID IDs across accounts/profile/address/order/cart and Vietnam two-level
    address fields: province/city and ward/commune only.'
servers:
- url: http://localhost:8080/api/v1
  description: Local Spring Boot API
- url: https://api.naherbs.vn/api/v1
  description: Production API placeholder
...
paths:
...
  /products:
    get:
      tags:
      - Public Products
      security: []
      summary: List published products
      parameters:
      - in: query
        name: keyword
        schema:
          type: string
      - in: query
        name: categorySlugs
        schema:
          type: array
          items:
            type: string
        explode: true
      - in: query
        name: minPrice
        schema:
          type: number
      - in: query
        name: maxPrice
        schema:
          type: number
      - in: query
        name: inStockOnly
        schema:
          type: boolean
      - in: query
        name: sort
        schema:
          type: string
          enum:
          - latest
          - price_asc
          - price_desc
      - in: query
        name: page
        schema:
          type: integer
          default: 0
      - in: query
        name: size
        schema:
          type: integer
          default: 12
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/ProductPage'
  /products/{slug}:
    get:
      tags:
      - Public Products
      security: []
      summary: Get product detail by slug
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/ProductDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/products:
    get:
      tags:
      - Admin Products
      summary: List products for admin
      parameters:
      - in: query
        name: keyword
        schema:
          type: string
      - in: query
        name: status
        schema:
          $ref: '#/components/schemas/ProductStatus'
      - in: query
        name: page
        schema:
          type: integer
          default: 0
      - in: query
        name: size
        schema:
          type: integer
          default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/ProductPage'
    post:
      tags:
      - Admin Products
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertProductRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/ProductDetail'
  /admin/products/{productId}:
    get:
      tags:
      - Admin Products
      summary: Get product details
      parameters:
      - in: path
        name: productId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Product details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/ProductDetail'
    put:
      tags:
      - Admin Products
      summary: Update product
      parameters:
      - in: path
        name: productId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertProductRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
    delete:
      tags:
      - Admin Products
      summary: Archive product
      parameters:
      - in: path
        name: productId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          $ref: '#/components/responses/Ok'
...
  /admin/product-skus/{skuId}:
    put:
      tags:
      - Admin Products
      summary: Update SKU
      parameters:
      - in: path
        name: skuId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertProductSkuRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /admin/product-skus/{skuId}/stock:
    patch:
      tags:
      - Admin Products
      summary: Update SKU stock
      parameters:
      - in: path
        name: skuId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStockRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
...
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ApiError'
...
    ApiError:
      type: object
      properties:
        field:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        message:
          type: string
    ProductStatus:
      type: string
      enum:
      - DRAFT
      - PUBLISHED
      - HIDDEN
      - ARCHIVED
    SkuStatus:
      type: string
      enum:
      - ACTIVE
      - HIDDEN
      - OUT_OF_STOCK
    StockStatus:
      type: string
      enum:
      - IN_STOCK
      - LOW_STOCK
      - OUT_OF_STOCK
...
    ProductCategorySummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
    ProductSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        shortDescription:
          type: string
          nullable: true
        minSalePrice:
          type: number
        maxSalePrice:
          type: number
        stockStatus:
          $ref: '#/components/schemas/StockStatus'
        skuCode:
          type: string
          nullable: true
        categoryName:
          type: string
          nullable: true
        categorySlug:
          type: string
          nullable: true
        skuCount:
          type: integer
        totalStockQuantity:
          type: integer
        status:
          $ref: '#/components/schemas/ProductStatus'
    ProductPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductSummary'
        page:
          type: integer
        size:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
    ProductDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        category:
          $ref: '#/components/schemas/ProductCategorySummary'
        shortDescription:
          type: string
        detailDescription:
          type: string
        usageInstruction:
          type: string
          nullable: true
        safetyNote:
          type: string
          nullable: true
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
        benefits:
          type: string
          nullable: true
        preservationInstruction:
          type: string
          nullable: true
        primaryKeyword:
          type: string
          nullable: true
        isFeatured:
          type: boolean
        displayOrder:
          type: integer
          format: int32
        versions:
          type: array
          items:
            $ref: '#/components/schemas/ProductVersion'
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImage'
        relatedProducts:
          type: array
          items:
            $ref: '#/components/schemas/ProductSummary'
    ProductVersion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        displayOrder:
          type: integer
        skus:
          type: array
          items:
            $ref: '#/components/schemas/ProductSku'
    ProductSku:
      type: object
      properties:
        id:
          type: string
          format: uuid
        skuCode:
          type: string
        name:
          type: string
        color:
          type: string
          nullable: true
        scent:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
        originalPrice:
          type: number
          nullable: true
        salePrice:
          type: number
        stockQuantity:
          type: integer
        stockStatus:
          $ref: '#/components/schemas/StockStatus'
        status:
          $ref: '#/components/schemas/SkuStatus'
        thumbnailUrl:
          type: string
          nullable: true
    ProductImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        altText:
          type: string
          nullable: true
        isThumbnail:
          type: boolean
...
    UpsertProductRequest:
      type: object
      required:
      - name
      - slug
      - status
      properties:
        categoryId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        slug:
          type: string
        shortDescription:
          type: string
          nullable: true
        detailDescription:
          type: string
          nullable: true
        usageInstruction:
          type: string
          nullable: true
        safetyNote:
          type: string
          nullable: true
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ProductStatus'
        benefits:
          type: string
          nullable: true
        preservationInstruction:
          type: string
          nullable: true
        primaryKeyword:
          type: string
          nullable: true
        isFeatured:
          type: boolean
        displayOrder:
          type: integer
          format: int32
        images:
          type: array
          items:
            $ref: '#/components/schemas/UpsertProductImageRequest'
          nullable: true
    UpsertProductImageRequest:
      type: object
      properties:
        mediaId:
          type: string
          format: uuid
          nullable: true
        url:
          type: string
        altText:
          type: string
          nullable: true
        isThumbnail:
          type: boolean
        displayOrder:
          type: integer
          format: int32
          nullable: true
...
    UpsertProductSkuRequest:
      type: object
      required:
      - versionId
      - skuCode
      - name
      - salePrice
      - stockQuantity
      - status
      properties:
        versionId:
          type: string
          format: uuid
        skuCode:
          type: string
        name:
          type: string
        color:
          type: string
          nullable: true
        scent:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
        originalPrice:
          type: number
          nullable: true
        salePrice:
          type: number
        stockQuantity:
          type: integer
        status:
          $ref: '#/components/schemas/SkuStatus'
    UpdateStockRequest:
      type: object
      required:
      - stockQuantity
      properties:
        stockQuantity:
          type: integer
          minimum: 0
        note:
          type: string
          nullable: true
...
security:
- bearerAuth: []