# naherb-web/src/services/generated/model/productDetail.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * NaHerbs Website MVP API
 * API contract for naherb-api (Spring Boot) used by naherb-web (Next.js). Updated for UUID IDs across accounts/profile/address/order/cart and Vietnam two-level address fields: province/city and ward/commune only.
 * OpenAPI spec version: 2.2.0
 */
import type { ProductCategorySummary } from './productCategorySummary';
import type { ProductVersion } from './productVersion';
import type { ProductImage } from './productImage';
import type { ProductSummary } from './productSummary';

export interface ProductDetail {
  id?: string;
  name?: string;
  slug?: string;
  category?: ProductCategorySummary;
  shortDescription?: string;
  detailDescription?: string;
  /** @nullable */
  usageInstruction?: string | null;
  /** @nullable */
  safetyNote?: string | null;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  /** @nullable */
  benefits?: string | null;
  /** @nullable */
  preservationInstruction?: string | null;
  /** @nullable */
  primaryKeyword?: string | null;
  isFeatured?: boolean;
  displayOrder?: number;
  versions?: ProductVersion[];
  images?: ProductImage[];
  relatedProducts?: ProductSummary[];
}