# naherb-web/src/services/generated/model/productSummary.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * NaHerbs Website MVP API
 * API contract for naherb-api (Spring Boot) used by naherb-web (Next.js). Updated for UUID IDs across accounts/profile/address/order/cart and Vietnam two-level address fields: province/city and ward/commune only.
 * OpenAPI spec version: 2.2.0
 */
import type { StockStatus } from './stockStatus';
import type { ProductStatus } from './productStatus';

export interface ProductSummary {
  id?: string;
  name?: string;
  slug?: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  shortDescription?: string | null;
  originalPrice?: number;
  minSalePrice?: number;
  maxSalePrice?: number;
  stockStatus?: StockStatus;
  /** @nullable */
  skuCode?: string | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  categorySlug?: string | null;
  skuCount?: number;
  totalStockQuantity?: number;
  status?: ProductStatus;
}