# naherb-web/src/services/generated/model/upsertProductSkuRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * NaHerbs Website MVP API
 * API contract for naherb-api (Spring Boot) used by naherb-web (Next.js). Updated for UUID IDs across accounts/profile/address/order/cart and Vietnam two-level address fields: province/city and ward/commune only.
 * OpenAPI spec version: 2.2.0
 */
import type { SkuStatus } from './skuStatus';

export interface UpsertProductSkuRequest {
  versionId: string;
  skuCode: string;
  name: string;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  scent?: string | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  originalPrice?: number | null;
  salePrice: number;
  stockQuantity: number;
  status: SkuStatus;
  /** @nullable */
  thumbnailMediaId?: string | null;
}